requirements.txt
atheris
PyYAML
markdown
tabulate
rich
